feat : 댓글삭제, 게임삭제 confirm 추가

+++ frontend/src/components/comments/comment/Comment.jsx
@@ -5,6 +5,7 @@ import timeAgo from '../../../utils/timeAgo';
 import useBestCommentLike from '../../../hooks/queries/useBestCommentLike';
 import useCommentLike from '../../../hooks/queries/useCommentLike';
 import useDeleteComment from '../../../hooks/queries/useDeleteComment';
+import { toast } from 'react-toastify';
 
 export default function Comment(props) {
   const { mutate: bestCommentLikeMutate } = useBestCommentLike(props.commentId, props.itemId);
@@ -21,7 +22,9 @@ export default function Comment(props) {
   };
 
   const handleDeleteComment = () => {
-    deleteCommentMutate();
+    if (window.confirm('삭제하시겠습니까?')) {
+      deleteCommentMutate();
+    }
   };
 
   return (
